requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.35.0
sqlalchemy==2.0.40
pydantic-settings==2.8.0
pwdlib==0.2.1
PyJWT==2.10.1
python-multipart==0.0.20
streamlit==1.44.1
requests==2.32.3
pwdlib[argon2]